requirements.txt
streamlit
numpy
matplotlib
transformers
torch
flair
librosa
streamlit-audiorecorder